# openapi.json
{
  "components": {
    "schemas": {
...
      "MoneyOut": {
        "properties": {
          "amount": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Amount",
            "type": "string"
          },
          "currency": {
            "title": "Currency",
            "type": "string"
          }
        },
        "required": [
          "amount",
          "currency"
        ],
        "title": "MoneyOut",
        "type": "object"
      },
      "OrderLineIn": {
        "properties": {
          "currency": {
            "pattern": "^[A-Z]{3}$",
            "title": "Currency",
            "type": "string"
          },
          "quantity": {
            "exclusiveMinimum": 0.0,
            "maximum": 2147483647.0,
            "title": "Quantity",
            "type": "integer"
          },
          "sku": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Sku",
            "type": "string"
          },
          "unit_amount": {
            "anyOf": [
              {
                "maximum": 999999999999.99,
                "minimum": 0.0,
                "multipleOf": 0.01,
                "type": "number"
              },
              {
                "pattern": "^(0|[1-9][0-9]{0,11})(\\.[0-9]{1,2})?$",
                "type": "string"
              }
            ],
            "title": "Unit Amount"
          }
        },
        "required": [
          "sku",
          "quantity",
          "unit_amount",
          "currency"
        ],
        "title": "OrderLineIn",
        "type": "object"
      },
      "OrderLineOut": {
        "properties": {
          "quantity": {
            "title": "Quantity",
            "type": "integer"
          },
          "sku": {
            "title": "Sku",
            "type": "string"
          },
          "subtotal": {
            "$ref": "#/components/schemas/MoneyOut"
          },
          "unit_price": {
            "$ref": "#/components/schemas/MoneyOut"
          }
        },
        "required": [
          "sku",
          "quantity",
          "unit_price",
          "subtotal"
        ],
        "title": "OrderLineOut",
        "type": "object"
      },
      "OrderResponse": {
        "properties": {
          "customer_id": {
            "format": "uuid",
            "title": "Customer Id",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "lines": {
            "items": {
              "$ref": "#/components/schemas/OrderLineOut"
            },
            "title": "Lines",
            "type": "array"
          },
          "total": {
            "$ref": "#/components/schemas/MoneyOut"
          }
        },
        "required": [
          "id",
          "customer_id",
          "lines",
          "total"
        ],
        "title": "OrderResponse",
        "type": "object"
      },
      "PlaceOrderRequest": {
        "properties": {
          "customer_id": {
            "format": "uuid",
            "title": "Customer Id",
            "type": "string"
          },
          "lines": {
            "items": {
              "$ref": "#/components/schemas/OrderLineIn"
            },
            "minItems": 1,
            "title": "Lines",
            "type": "array"
          }
        },
        "required": [
          "customer_id",
          "lines"
        ],
        "title": "PlaceOrderRequest",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "pyfr-m8-verify",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/orders": {
      "post": {
        "operationId": "place_order_api_v1_orders_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlaceOrderRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "properties": {
                    "detail": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ],
                      "title": "Detail"
                    },
                    "instance": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ],
                      "title": "Instance"
                    },
                    "status": {
                      "title": "Status",
                      "type": "integer"
                    },
                    "title": {
                      "title": "Title",
                      "type": "string"
                    },
                    "type": {
                      "default": "about:blank",
                      "title": "Type",
                      "type": "string"
                    }
                  },
                  "required": [
                    "title",
                    "status"
                  ],
                  "title": "ProblemDetail",
                  "type": "object"
                }
              }
            },
            "description": "Malformed request"
          },
          "402": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "properties": {
                    "detail": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ],
                      "title": "Detail"
                    },
                    "instance": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ],
                      "title": "Instance"
                    },
                    "status": {
                      "title": "Status",
                      "type": "integer"
                    },
                    "title": {
                      "title": "Title",
                      "type": "string"
                    },
                    "type": {
                      "default": "about:blank",
                      "title": "Type",
                      "type": "string"
                    }
                  },
                  "required": [
                    "title",
                    "status"
                  ],
                  "title": "ProblemDetail",
                  "type": "object"
                }
              }
            },
            "description": "Payment declined"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "properties": {
                    "detail": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ],
                      "title": "Detail"
                    },
                    "instance": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ],
                      "title": "Instance"
                    },
                    "status": {
                      "title": "Status",
                      "type": "integer"
                    },
                    "title": {
                      "title": "Title",
                      "type": "string"
                    },
                    "type": {
                      "default": "about:blank",
                      "title": "Type",
                      "type": "string"
                    }
                  },
                  "required": [
                    "title",
                    "status"
                  ],
                  "title": "ProblemDetail",
                  "type": "object"
                }
              }
            },
            "description": "No such resource"
          },
          "405": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "properties": {
                    "detail": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ],
                      "title": "Detail"
                    },
                    "instance": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ],
                      "title": "Instance"
                    },
                    "status": {
                      "title": "Status",
                      "type": "integer"
                    },
                    "title": {
                      "title": "Title",
                      "type": "string"
                    },
                    "type": {
                      "default": "about:blank",
                      "title": "Type",
                      "type": "string"
                    }
                  },
                  "required": [
                    "title",
                    "status"
                  ],
                  "title": "ProblemDetail",
                  "type": "object"
                }
              }
            },
            "description": "Method not allowed"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "properties": {
                    "detail": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ],
                      "title": "Detail"
                    },
                    "instance": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ],
                      "title": "Instance"
                    },
                    "status": {
                      "title": "Status",
                      "type": "integer"
                    },
                    "title": {
                      "title": "Title",
                      "type": "string"
                    },
                    "type": {
                      "default": "about:blank",
                      "title": "Type",
                      "type": "string"
                    }
                  },
                  "required": [
                    "title",
                    "status"
                  ],
                  "title": "ProblemDetail",
                  "type": "object"
                }
              }
            },
            "description": "Request validation failed"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "properties": {
                    "detail": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ],
                      "title": "Detail"
                    },
                    "instance": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ],
                      "title": "Instance"
                    },
                    "status": {
                      "title": "Status",
                      "type": "integer"
                    },
                    "title": {
                      "title": "Title",
                      "type": "string"
                    },
                    "type": {
                      "default": "about:blank",
                      "title": "Type",
                      "type": "string"
                    }
                  },
                  "required": [
                    "title",
                    "status"
                  ],
                  "title": "ProblemDetail",
                  "type": "object"
                }
              }
            },
            "description": "Internal server error"
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "properties": {
                    "detail": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ],
                      "title": "Detail"
                    },
                    "instance": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ],
                      "title": "Instance"
                    },
                    "status": {
                      "title": "Status",
                      "type": "integer"
                    },
                    "title": {
                      "title": "Title",
                      "type": "string"
                    },
                    "type": {
                      "default": "about:blank",
                      "title": "Type",
                      "type": "string"
                    }
                  },
                  "required": [
                    "title",
                    "status"
                  ],
                  "title": "ProblemDetail",
                  "type": "object"
                }
              }
            },
            "description": "Payment provider unavailable"
          }
        },
        "summary": "Place Order",
        "tags": [
          "orders"
        ]
      }
    },
    "/api/v1/orders/{order_id}": {
      "get": {
        "operationId": "get_order_api_v1_orders__order_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "order_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Order Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "properties": {
                    "detail": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ],
                      "title": "Detail"
                    },
                    "instance": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ],
                      "title": "Instance"
                    },
                    "status": {
                      "title": "Status",
                      "type": "integer"
                    },
                    "title": {
                      "title": "Title",
                      "type": "string"
                    },
                    "type": {
                      "default": "about:blank",
                      "title": "Type",
                      "type": "string"
                    }
                  },
                  "required": [
                    "title",
                    "status"
                  ],
                  "title": "ProblemDetail",
                  "type": "object"
                }
              }
            },
            "description": "Malformed request"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "properties": {
                    "detail": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ],
                      "title": "Detail"
                    },
                    "instance": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ],
                      "title": "Instance"
                    },
                    "status": {
                      "title": "Status",
                      "type": "integer"
                    },
                    "title": {
                      "title": "Title",
                      "type": "string"
                    },
                    "type": {
                      "default": "about:blank",
                      "title": "Type",
                      "type": "string"
                    }
                  },
                  "required": [
                    "title",
                    "status"
                  ],
                  "title": "ProblemDetail",
                  "type": "object"
                }
              }
            },
            "description": "Order not found"
          },
          "405": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "properties": {
                    "detail": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ],
                      "title": "Detail"
                    },
                    "instance": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ],
                      "title": "Instance"
                    },
                    "status": {
                      "title": "Status",
                      "type": "integer"
                    },
                    "title": {
                      "title": "Title",
                      "type": "string"
                    },
                    "type": {
                      "default": "about:blank",
                      "title": "Type",
                      "type": "string"
                    }
                  },
                  "required": [
                    "title",
                    "status"
                  ],
                  "title": "ProblemDetail",
                  "type": "object"
                }
              }
            },
            "description": "Method not allowed"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "properties": {
                    "detail": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ],
                      "title": "Detail"
                    },
                    "instance": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ],
                      "title": "Instance"
                    },
                    "status": {
                      "title": "Status",
                      "type": "integer"
                    },
                    "title": {
                      "title": "Title",
                      "type": "string"
                    },
                    "type": {
                      "default": "about:blank",
                      "title": "Type",
                      "type": "string"
                    }
                  },
                  "required": [
                    "title",
                    "status"
                  ],
                  "title": "ProblemDetail",
                  "type": "object"
                }
              }
            },
            "description": "Request validation failed"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "properties": {
                    "detail": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ],
                      "title": "Detail"
                    },
                    "instance": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ],
                      "title": "Instance"
                    },
                    "status": {
                      "title": "Status",
                      "type": "integer"
                    },
                    "title": {
                      "title": "Title",
                      "type": "string"
                    },
                    "type": {
                      "default": "about:blank",
                      "title": "Type",
                      "type": "string"
                    }
                  },
                  "required": [
                    "title",
                    "status"
                  ],
                  "title": "ProblemDetail",
                  "type": "object"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "summary": "Get Order",
        "tags": [
          "orders"
        ]
      }
    },
...
  }
}